Move `graphpublisher` tests into `graph-publisher`
Because it is only used by these tests, `TestMeasurementAggregatedMock` was
also moved with the Graph Publisher test classes.

I've consider moving it into the `talaiot-test-utils` module in a mock
package, but it was not possible due to the dependency on the classes
from `com.cdsap.talaiot.entities` from the `talaiot` module.
