+++ library/plugins/graph/graph-publisher/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("kotlinLib")
+    `kotlin-dsl`
 }
 
 dependencies {
