+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/GexfPublisherTest.kt
@@ -1,8 +1,7 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesGexf
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesGexf
 import com.cdsap.talaiot.logger.LogTracker
-import com.cdsap.talaiot.mock.TaskMeasurementAggregatedMock
 import com.cdsap.talaiot.utils.TestExecutor
 import com.cdsap.talaiot.writer.FileWriter
 import com.nhaarman.mockitokotlin2.argThat
