+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/GraphPublisherFactoryTest.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.publisher.Publisher
