+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/HtmlPublisherTest.kt
@@ -1,9 +1,9 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesHtml
 import com.cdsap.talaiot.logger.LogTracker
-import com.cdsap.talaiot.mock.TaskMeasurementAggregatedMock.taskMeasurementAggregated
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml.LEGEND_HEADER
+import com.cdsap.talaiot.publisher.graph.TaskMeasurementAggregatedMock.taskMeasurementAggregated
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesHtml.LEGEND_HEADER
 import com.cdsap.talaiot.utils.TestExecutor
 import com.cdsap.talaiot.writer.FileWriter
 import com.nhaarman.mockitokotlin2.argThat
